Небольшой рефакторинг кода
