+++ app/src/main/java/com/evgeny_m/messenger3/utils/navDraver.kt
@@ -2,13 +2,11 @@ package com.evgeny_m.messenger3.utils
 
 import androidx.appcompat.widget.Toolbar
 import androidx.drawerlayout.widget.DrawerLayout
-import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupWithNavController
 import com.evgeny_m.messenger3.MainActivity
 import com.evgeny_m.messenger3.R
-import com.evgeny_m.messenger3.fragments.main.navHeader.NavHeaderFragment
 
 lateinit var drawerLayout: DrawerLayout
 lateinit var APP: MainActivity
@@ -21,14 +19,14 @@ fun MainActivity.initNavDrawer() {
     navView.setupWithNavController(navController)
 }
 
-fun Fragment.initMainNavigationButton(toolbar: Toolbar) {
+fun initMainNavigationButton(toolbar: Toolbar) {
     toolbar.setNavigationIcon(R.drawable.icon_gamburger_button)
     toolbar.setNavigationOnClickListener {
         drawerLayout.open()
     }
 }
 
-fun Fragment.initBackButton(toolbar: Toolbar) {
+fun initBackButton(toolbar: Toolbar) {
     toolbar.setNavigationIcon(R.drawable.icon_back_button)
     drawerLayout.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED)
     toolbar.setNavigationOnClickListener {
